// src/Tawtheef.Websites/operations-web/src/app/pages/user/employee/job-management/models/job-tab-review-note-response.ts
import { GUID } from "../../../../../shared/types/guid.type";
import { JobTabStatus } from "../enums/job-tab-status";
import { JobTabType } from "../enums/job-tab-type";

export interface JobTabReviewNoteResponse {
  id: GUID;
  tab: JobTabType;
  note: string;
  tabStatus: JobTabStatus;
  isResolved: boolean;
}